+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.7.0]
 
 - feat: Allow manual editing of the number of questions in the AI generation dialog via a text field.
+- feat: Added "Content Mode" selection (Theory, Exercises,Mixed) to AI generation to refine results and avoid unwanted generic exercises.
 - fix: Prevent collapsing a `QuestionPreviewCard` when tapping collapsible content.
 - feat: Allow execute a quiz with only-selected questions.